% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{boot_eff.gamlss}
\alias{boot_eff.gamlss}
\title{Bootstraps (percentile) confidence intervals for effects from a GAMLSS model.}
\usage{
boot_eff.gamlss(
  obj,
  data,
  R = 1000,
  what = c("mu", "sigma", "nu", "tau"),
  terms,
  progress = TRUE,
  ...
)
}
\arguments{
\item{obj}{A fitted `gamlss` object.}

\item{data}{The original data used to fit `obj`.}

\item{R}{Number of bootstrap iterations}

\item{what}{Which moment of the distribution to predict - only one allowed at a time.}

\item{terms}{Named list of variable names and values used to generate predictions. All
possible combinations of variables provided will be expanded.}

\item{progress}{Logical indicating whether a progress bar should be printed.}

\item{...}{Currently not implemented.}
}
\description{
Bootstraps (percentile) confidence intervals for effects from a GAMLSS model.
}
\details{
Uses nonparametric bootstrapping to calculate confidence intervals for effects
plots.  As the function refits the model for every bootstrap iteration, computation times
could be long. One of the main benefits of this approach is it also essentially takes into
account variability in the smoothing parameters of the penalized splines.  Use `term.plot` from
the `gamlss` package as a quick alternative.
}
