% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{brief.gamlss}
\alias{brief.gamlss}
\title{Brief summary of GAMLSS objects}
\usage{
\method{brief}{gamlss}(obj, ...)
}
\arguments{
\item{obj}{A `gamlss` object}

\item{...}{currently not implemented}
}
\description{
Brief summary of GAMLSS objects
}
