% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{brief}
\alias{brief}
\title{Brief summary of gamlss objects}
\usage{
brief(obj, ...)
}
\arguments{
\item{obj}{A `gamlss` object}

\item{...}{currently not implemented}
}
\description{
Brief summary of gamlss objects
}
