% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{bwn}
\alias{bwn}
\title{Between Group Transformation Function}
\usage{
bwn(x, id, means = NULL, ...)
}
\arguments{
\item{x}{A variable whose within transformation you want to calculate. This should be a vector of values, not a variable name.}

\item{id}{The grouping variable within which the transformation be performed.}

\item{means}{An optional vector of matrix of means depending on the class of `x`.}

\item{...}{Currently not implemented.}
}
\description{
Performs between transformation on `x` within the groups of `id`.  If `x` is numeric, 
then `means` should either be `NULL` or a named vector of means with the names of the
elements have corresponding elements in `id`.  If `x` is a factor, then `means` should
be an `id` by levels of `x` matrix giving the proportion of observations in each category
of `x` for each different value of `id`.  If `means` is `NULL`, it will be calculated 
internally and then passed to the function.
}
