% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{bwnt.numeric}
\alias{bwnt.numeric}
\title{Numeric method for `bwnt`}
\usage{
\method{wint}{numeric}(x, id, means = NULL, ...)
}
\arguments{
\item{x}{A variable whose between transformation you want to calculate. This should be a vector of values, not a variable name.}

\item{id}{The grouping variable between which the transformation be performed.}

\item{means}{An optional vector of matrix of means depending on the class of `x`.}

\item{...}{Currently not implemented.}
}
\description{
Numeric method for `bwnt`
}
