% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{calc_ymax}
\alias{calc_ymax}
\title{Calculates Values from a Plot}
\usage{
calc_ymax(plot, scale = 0.3)
}
\arguments{
\item{plot}{A `ggplot`}

\item{scale}{The fraction of the y-axis to identify.  For example, if `scale = .3`, then the function will return a value 30 percent of the way up the y-axis.}
}
\description{
Calculates Values from a Plot
}
