% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{cr.gamlss}
\alias{cr.gamlss}
\title{Create CR Plots for GAMLSS objects}
\usage{
cr.gamlss(
  obj,
  term,
  plot = TRUE,
  points = TRUE,
  what = c("mu", "sigma", "nu", "tau"),
  ...
)
}
\arguments{
\item{obj}{A `gamlss` object}

\item{term}{Name of the variable for the plot}

\item{plot}{Logical indicating whether the plot should be returned (`TRUE`) or just the data (`FALSE`)}

\item{points}{Should the points be plotted too (`TRUE`) or just the lines (`FALSE`)}

\item{what}{Which moment of the distribution should be considered}

\item{...}{Other arguments, currently not implemented}
}
\description{
Create CR Plots for GAMLSS objects
}
