% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{-.gg}
\alias{-.gg}
\title{Plot layer behind existing layers in ggplot}
\usage{
\method{-}{gg}(plot, layer)
}
\arguments{
\item{plot}{A `ggplot`}

\item{layer}{A layer to be added to the plot}
}
\description{
Plot layer behind existing layers in ggplot
}
