% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{IC_delta}
\alias{IC_delta}
\title{Calculates Information Criterion Delta Values}
\usage{
IC_delta(...)
}
\arguments{
\item{...}{Models whose IC values will be calculated}
}
\description{
Calculates Information Criterion Delta Values
}
