% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{make_between_data}
\alias{make_between_data}
\title{Make between and within group data}
\usage{
make_between_data(formula, data, id)
}
\arguments{
\item{formula}{Formula with DV and IVs to be transformed.}

\item{data}{A data frame the contains the variables in `formula`.}

\item{id}{Character vector giving the names of the group ID variable(s).}
}
\description{
Makes both between and within transformed variables to use in a hierarchical model
}
