% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{make_post_data}
\alias{make_post_data}
\title{Make Data for MICE Posterior Predictive Distributions}
\usage{
make_post_data(data, ...)
}
\arguments{
\item{data}{The original data that were used in the `mice()` call.}

\item{...}{Currently not implemented.}
}
\description{
Makes data that will permit the generation of the posterior predictive
distribution using `mice()` from the `mice` package.  The function replicates
the original input data `k` times, each time setting one of the variables 
to completely missing such that all its values will be imputed from the imputation 
model.  See the help for `process_post()` for managing the data after imputation.
}
