% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{makeFakeData}
\alias{makeFakeData}
\title{Make Hypothetical Prediction Data for mlogit models}
\usage{
makeFakeData(obj, data, change)
}
\arguments{
\item{obj}{An object of class `mlogit`}

\item{data}{The data frame used to fit `obj`}

\item{change}{A named list with a single element whose values are the values over which the named variable will be varied}
}
\description{
Make Hypothetical Prediction Data for mlogit models
}
