% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{makepredictcall.bwn}
\alias{makepredictcall.bwn}
\title{Makepredictcall method for bwn}
\usage{
\method{makepredictcall}{bwn}(var, call)
}
\arguments{
\item{var}{A variable}

\item{call}{A term in the formula, as a call.}
}
\description{
Makepredictcall method for bwn
}
