% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{makepredictcall.win}
\alias{makepredictcall.win}
\title{Makepredictcall method for win}
\usage{
\method{makepredictcall}{win}(var, call)
}
\arguments{
\item{var}{A variable}

\item{call}{A term in the formula, as a call.}
}
\description{
Makepredictcall method for win
}
