% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{predict.bwn}
\alias{predict.bwn}
\title{Predict method for bwn}
\usage{
\method{predict}{bwn}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class `bwn`}

\item{newdata}{An optional data frame with which to generate the predictions}

\item{...}{Other arguments to be passed down}
}
\description{
Predict method for bwn
}
