% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{predict.win}
\alias{predict.win}
\title{Predict method for win}
\usage{
\method{predict}{win}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class `win`}

\item{newdata}{An optional data frame with which to generate the predictions}
}
\description{
Predict method for win
}
