% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{print.brief}
\alias{print.brief}
\title{Print method for brief summaries}
\usage{
\method{print}{brief}(x, ...)
}
\arguments{
\item{x}{an object of class `brief`}

\item{...}{currently not implemented}
}
\description{
Print method for brief summaries
}
