% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{process_post}
\alias{process_post}
\title{Process Posterior Predictive MICE Distributions}
\usage{
process_post(imps, data, ...)
}
\arguments{
\item{imps}{A `mice` object that was estimated on data that were generated
using `make_post_data()`.}

\item{data}{The data used in the original call to `mice`.}

\item{...}{Currently not implemented.}
}
\description{
Processes the posterior predictive distributions generated by `mice()` 
using the output from `make_post_dat()`.
}
