% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{restplot}
\alias{restplot}
\title{Calculates the Rest Plot for the Summated Rating Model}
\usage{
restplot(X, ...)
}
\arguments{
\item{X}{A matrix that was subjected to a summated rating analysis.}

\item{...}{Arguments passed to `method.args` in `geom_smooth()`.}
}
\description{
The rest plot plots the scale constructed using the "rest" of the data against a single excluded variable.
}
