% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{tidy.comparisons}
\alias{tidy.comparisons}
\title{Tidy method for `comparisons` from `marginaleffects` package}
\usage{
\method{tidy}{comparisons}(x, ...)
}
\arguments{
\item{x}{An object of class `comparisons`}

\item{...}{currently not implemented}
}
\description{
Tidy method for `comparisons` from `marginaleffects` package
}
