% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{tp_data}
\alias{tp_data}
\title{Generate Effect Plot Data}
\usage{
tp_data(obj, data, what = c("mu", "sigma", "nu", "tau"), ...)
}
\arguments{
\item{obj}{A `gamlss` object}

\item{data}{The data frame used to fit `obj`.}

\item{what}{The moment of the data to be predicted}

\item{...}{Currently not implemented}
}
\description{
Generate linear predictor and standard error values from GAMLSS models
}
