% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{win_eff}
\alias{win_eff}
\title{Calculates Effects for Within-transformed Variables}
\usage{
win_eff(obj, vbl, idvar, data, nvals = 25, usevals = NULL, ...)
}
\arguments{
\item{obj}{A `lmerMod` or `glmerMod` object - estimated with `lmer` or `glmer` from 
the `lme4` package.}

\item{vbl}{A character string giving the name of a variable whose effects are to be calculated.}

\item{idvar}{A character string giving the name of the grouping variable.}

\item{data}{The data frame used to fit the original model.}

\item{nvals}{The number of values to use when varying a numeric variable.  The 
values used in the prediction will be a sequence of `nvals` evenly spaced from the 
minimum to the maximum of `vbl`.  Disregarded if `usevals` is specified.}

\item{usevals}{The values to use in generating predictions for numeric variables.}

\item{...}{Other arguments that get passed down to `lme4:::predict.merMod()`.}
}
\description{
Calculates predicted values for within-transformed variables holding
all other variables constant at median/modal values (depending on whether the
variable is numeric or a factor).  All variables must be either factors or 
numeric, the function will fail if character variables are used.  Variables
with between transformations will be held constant at each group's mean.  Currently
the only option is to make the desired effect for each different random group.  
Currently this only works with two-levels models that have a single random effect. 
The function returns predictions on the link scale.
}
