% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uwo9592.R
\name{wint}
\alias{wint}
\title{Performs within transformation}
\usage{
wint(x, id, means = NULL, ...)
}
\arguments{
\item{x}{A variable whose within transformation you want to calculate. This should be a vector of values, not a variable name.}

\item{id}{The grouping variable within which the transformation be performed.}

\item{means}{An optional vector of matrix of means depending on the class of `x`.}

\item{...}{Currently not implemented.}
}
\description{
Performs within transformation
}
